Rename ggrs_stage module to schedule_systems
